+++ backend/src/main/java/br/com/carreselling/usecase/vehicle/list/contract/VehicleListItem.java
@@ -15,5 +15,5 @@ public record VehicleListItem(UUID id,
                               BigDecimal servicesTotal,
                               BigDecimal totalCost,
                               String assignedPartnerName,
-                              Long yardDays) {
+                              int yardDays) {
 }
